requirements.txt
asgiref==3.2.10
# boto3==1.17.36  # Removed for local-only
# botocore==1.20.36  # Removed for local-only
certifi==2020.12.5
chardet==4.0.0
Django==3.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.5
django-session-timeout==0.1.0
# django-storages==1.11.1  # Removed for local-only
idna==2.10
# jmespath==0.10.0  # Removed for local-only
Pillow==10.0.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.1
requests==2.25.1
# s3transfer==0.3.6  # Removed for local-only
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.3
